Extract changes into individual files.
This will make the changes easier to maintain and supplement.
