+++ src/changes/mod.rs
@@ -1,0 +1,11 @@
+mod append_ini_entry;
+mod change;
+mod rename_file;
+mod replace_in_file;
+mod set_ini_entry;
+
+pub use append_ini_entry::*;
+pub use change::*;
+pub use rename_file::*;
+pub use replace_in_file::*;
+pub use set_ini_entry::*;
